+++ server/src/protocol/test.proto.ts
@@ -1,9 +1,9 @@
-import { uint8 } from "./types";
+import { uint32 } from "./types";
 
 /**
  * @packer {id: 0xf1}
  */
 export interface TestPacket {
-    a: uint8;
-    b: uint8;
+    a: uint32;
+    b: uint32;
 }
